Update db-setup scripts and artifactory properties builder
Updated db-setup.sh and db-setup.js to accept environment variables for the root db user and password to use, and also to use env vars for the username and password to create. Also added a sleep so that it should not fail when it runs. ( may still fail... )

Updated the artifactory properties builder to work against, since it had become stale.